+++ backend/canisters/local_user_index/api/src/updates/bot_create_channel.rs
@@ -21,6 +21,7 @@ pub struct Args {
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum Response {
     Success(SuccessResult),
+    FailedAuthentication(String),
     InvalidRequest(String),
     NotAuthorized,
     Frozen,