+++ backend/canisters/local_user_index/api/src/updates/bot_send_message.rs
@@ -15,6 +15,7 @@ pub struct Args {
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum Response {
     Success(SuccessResult),
+    FailedAuthentication(String),
     InvalidRequest(String),
     NotAuthorized,
     Frozen,