+++ backend/canisters/local_user_index/impl/src/updates/mod.rs
@@ -1,4 +1,5 @@
 pub mod bot_create_channel;
+pub mod bot_delete_channel;
 pub mod bot_send_message;
 pub mod c2c_notify_low_balance;
 pub mod c2c_notify_user_events;